\name{datadyncouponbonds}
\alias{datadyncouponbonds}
\docType{data}
\title{German Government Bond Data Set}
\description{Dynamic German government coupon bond data set }

\usage{data(datadyncouponbonds)}


\note{
If you use your own data set, make sure that the structure is identical to the provided data sets.
Use the function \command{str()} to explore the data set.}

\keyword{datasets}
