\name{eurobonds}
\alias{eurobonds}
\docType{data}
\title{European Government Bonds}
\description{European government bonds.}

\usage{data(eurobonds)}

\details{
The data set \command{eurobonds} consists of German, Austrian, Italian and Hungarian government bonds.
}

\note{
If you use your own data set, make sure that the structure is identical to the provided data sets.
Use the function \command{str()} to explore the data set.}


\seealso{\command{\link{corpbonds}}, \command{\link{govbonds}}}

\examples{
data(corpbonds)
str(corpbonds) 
}

\keyword{datasets}
