\name{fwr_ns}
\alias{fwr_ns}

\title{Forward Rate Calculation according to Nelson/Siegel }
\description{Calculate forward rates according to the Nelson/Siegel approach.}
\usage{
fwr_ns(beta, m)
}

\arguments{
  \item{beta}{Parameter vector }
  \item{m}{Maturity vector }
}
%\details{}
\value{The function returns the calculated forward rate vector.}


%\references{  }
\author{ Robert Ferstl, Josef Hayden }
%\note{ }
\seealso{\code{\link{fwr_sv}}, \code{\link{forwardrates}}}
%\examples{}

\keyword{  }

