\name{fwr_sv}
\alias{fwr_sv}

\title{Forward Rate Calculation according to Svensson (1994).}
\description{
Calculate forward rates according to the Svensson. 
}
\usage{
fwr_sv(beta, m)
}

\arguments{
  \item{beta}{parameter vector.}
  \item{m}{maturity or vector of maturities. }
}
\details{The forward rates are calculated according to the following formula:
!!Formel !!


}
\value{Returns the a vector with the calculated forward rates.}


\references{Lars E.O. Svensson (1994): 
Estimating and Interpreting Forward Interest Rates: Sweden 1992 -1994. 
\emph{Technical Reports 4871, National Bureau of Economic Research}.   }
\author{ Robert Ferstl, Josef Hayden }
%\note{ }
\seealso{\code{\link{fwr_ns}}, \code{\link{forwardrates}}}
%\examples{}

\keyword{  }
