\name{get_constraints}
\alias{get_constraints}

\title{Constraints Selection
}
\description{Selection of the appropriate contraints for constrOptim()
}
\usage{
get_constraints(method, tauconstr)
}

\arguments{
  \item{method}{term structure estimation method 
  }
  \item{tauconstr}{constraints on tau parameters} 
}

