\name{get_grad_objfct_bonds}
\alias{get_grad_objfct_bonds}

\title{Gradient Selection Function }
\description{Selects the appropriate gradient of the objective function for a bond data set

}
\usage{
get_grad_objfct_bonds(method)
}

\arguments{
  \item{method}{term structure estimation method 

}
}
