\name{gi}
\alias{gi}

\title{ Cubic Functions }
\description{
  Calculation of the cubic functions according to an approach of
  J. Huston McCulloch (1971)
}
\usage{gi(t, T, i, s)}

\arguments{
  \item{t}{\code{t}}
  \item{T}{\code{T}}
  \item{i}{\code{i}}
  \item{s}{\code{s}}
}
\details{...}
\value{
 \item{g}{..g...}
 }
\references{ 

J. Huston McCulloch (1971):
Measuring the Term Structure of Interest Rates.
\emph{The Journal of Business, \bold{44} 19--31.} 

 }
\author{Rober Ferstl, Josef Hayden }
%\note{ }
%\seealso{  }
%\examples{e.g. example from the textbook }

\keyword{ }

