\name{govbonds}
\alias{govbonds}
\docType{data}
\title{European Government Bonds}
\description{European government bonds.}

\usage{data(govbonds)}

\details{
The data set \command{bonds} consists of German, Austrian and French government bonds.
}

\note{
If you use your own data set, make sure that the structure is identical to the provided data sets.
Use the function \command{str()} to explore the data set.}

\examples{
data(govbonds)
str(govbonds) 
}

\keyword{datasets}
