\name{grad_dl}
\alias{grad_dl}


\title{Gradient of the Diebold/Li Loss Function for Yields
}
\description{
Calculates the gradient of the Diebold/Li Loss Function for Yields
}



\usage{
grad_dl(beta, lambda, m, y)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{beta}{Parameter of the Diebold/Li spot rate function 
}
  \item{lambda}{constant parameter of the Diebold/Li spot rate function
}
  \item{m}{maturity vector
}
  \item{y}{yield vector
}
}
