\name{grad_sv}
\alias{grad_sv}

\title{Gradient of the Svensson Loss Function for Yields
}
\description{
Calculates the gradient of the Svensson Loss Function for Yields
}

\usage{
grad_sv(beta, m, y)
}



\arguments{
  \item{beta}{Parameter of the Svensson spot rate function.
}
  \item{m}{maturity vector
}
  \item{y}{yield vector
}
}