\name{impl_fwr}
\alias{impl_fwr}

\title{Implied forward rate calculation}
\description{Calculate the implied forward rates from given spot rates
}
\usage{
impl_fwr(m, s)
}


\arguments{
  \item{m}{Maturity vector}
  \item{s}{Spot rate vector}
}
%\details{}
\value{
The function returns the calculated forward rate vector.
}
%\references{  }
\author{ Robert Ferstl, Josef Hayden }
%\note{ }
\seealso{}
%\examples{}

\keyword{  }

