\name{maturity_range}
\Rdversion{1.1}
\alias{maturity_range}


\title{Restricting a Bond Dataset}

\description{The function restricts a bond data set to a specified maturity range.}

\usage{maturity_range(bonddata, lower, upper)}



\arguments{
  \item{bonddata}{bond data set.}
  \item{lower}{lower bound of maturity spectrum.}
  \item{upper}{upper bound of maturity spectrum.}
}

\note{Internal helper function.}