\name{objfct_asv_bonds}
\alias{objfct_asv_bonds}

\title{Adjusted Svensson Loss Function for Bonds
}
\description{Calculates the sum of the weighted squared price error.
}
\usage{
objfct_asv_bonds(beta, m, cf, w, p)
}

\arguments{
  \item{beta}{Paramter vector of the adjusted Svensson spot rate function 

}
  \item{m}{maturity matrix 
}
  \item{cf}{cashflow matrix
}
  \item{w}{weights
}
  \item{p}{price vector
}
}
