\name{objfct_asv_grid}
\alias{objfct_asv_grid}


\title{Adjusted Svensson Grid Loss Function for Yields}
\description{Calculates the sum of the squared yield error.}

\usage{
objfct_asv_grid(beta, tau,  m, y)
}

\arguments{
  \item{beta}{Beta parameters of adjusted Svensson spot rate function 

}
  \item{tau}{Tau parameters of adjusted Svensson spot rate function
}
  \item{m}{maturity vector}
  \item{y}{yield vector
}
}


