\name{objfct_ns_bonds}
\alias{objfct_ns_bonds}

\title{Nelson/Siegel Loss Function for Bonds

}
\description{Calculates the sum of the weighted squared price error.
}
\usage{
objfct_ns_bonds(beta, m, cf, w, p)
}

\arguments{
  \item{beta}{Parameter vector of the Nelson/Siegel spot rate function
}
  \item{m}{maturity matrix 
}
  \item{cf}{cashflow matrix
}
  \item{w}{weights
}
  \item{p}{price vector
}
}


