\name{objfct_ns_grid}
\alias{objfct_ns_grid}


\title{Nelson/Siegel Grid Loss Function for Yields}
\description{Calculates the sum of the squared yield error.}

\usage{
objfct_ns_grid(beta, tau,  m, y)
}

\arguments{
  \item{beta}{Beta parameters of the Nelson/Siegel spot rate function 

}
  \item{tau}{Tau parameter of Nelson/Siegel spot rate function
}
  \item{m}{maturity vector}
  \item{y}{yield vector
}
}