\name{objfct_sv_bonds}
\alias{objfct_sv_bonds}

\title{Svensson Loss Function for Bonds}
\description{Calculates the sum of the weighted squared price error. 
}
\usage{
objfct_sv_bonds(beta, m, cf, w, p)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{beta}{Parameter vector of the Svensson spot rate function.
}
  \item{m}{maturity matrix 
}
  \item{cf}{cashflow matrix
}
  \item{w}{weights
}
  \item{p}{price vector
}
}




