\name{objfct_sv_bonds_grid}
\alias{objfct_sv_bonds_grid}
\title{
Svensson grid loss function for bonds 
}
\description{
Svensson grid loss function for bonds 
}
\usage{
objfct_sv_bonds_grid(beta, tau, m, cf, w, p)
}

\arguments{
  \item{beta}{parameters}
  \item{tau}{parameters}
  \item{m}{maturities}
  \item{cf}{cash flows}
  \item{w}{weights}
  \item{p}{prices}
}

