\name{objfct_sv_grid}
\alias{objfct_sv_grid}


\title{Svensson Grid Loss Function for Bonds}
\description{Calculates the sum of the squared yield error.}

\usage{
objfct_sv_grid(beta, tau,  m, y)
}

\arguments{
  \item{beta}{Beta parameters of the Svensson spot rate function 

}
  \item{tau}{Tau parameters of the Svensson spot rate function
}
  \item{m}{maturity vector}
  \item{y}{yield vector
}
}