\name{param.dyntermstrc_nss}
\Rdversion{1.1}
\alias{param.dyntermstrc_nss}

\title{S3 Param Method}
\description{The function extracts the estimated term structure parameters from

 an object of the class \command{"dyntermstrc_nss"}.}
\usage{
\method{param}{dyntermstrc_nss}(x, ...)
}

\arguments{
  \item{x}{object of the class \command{"dyntermstrc_nss"}}
  \item{\dots}{further arguments}
}

\seealso{
 \code{\link{param}}
}
