\name{param.dyntermstrc_yields}
\Rdversion{1.1}
\alias{param.dyntermstrc_yields}

\title{S3 Param Method}
\description{The function extracts the estimated term structure parameters from 
 
an object of the class \command{"dyntermstrc_yields"}.}
\usage{
\method{param}{dyntermstrc_yields}(x, ...)
}

\arguments{
  \item{x}{object of the class \command{"dyntermstrc_yields"}}
  \item{\dots}{further arguments}
}

\seealso{
 \code{\link{param}}
}
