\name{plot.dyntermstrc_nss}
\Rdversion{1.1}
\alias{plot.dyntermstrc_nss}

\title{S3 Plot Method}
\description{Plot method for objects of the class
  \command{"dyntermstrc_nss"}. The method plots the estimated three-dimensional spot
rate curve.}
\usage{
\method{plot}{dyntermstrc_nss}(x, range = c(0, 20), ...)
}

\arguments{
  \item{x}{object of the class \command{"dyntermstrc_nss"}}
  \item{range}{maturity range, e.g., \command{c(0,20)} (default)}
  \item{\dots}{further arguments}
}
