\name{plot.dyntermstrc_yields}
\Rdversion{1.1}
\alias{plot.dyntermstrc_yields}

\title{S3 Plot Method
}
\description{Plot method for object of the class
  \command{"dyntermstrc_yields"}. The method plot the estimated
  three-dimensional spot rate curve.
}
\usage{
\method{plot}{dyntermstrc_yields}(x, ...)
}

\arguments{
  \item{x}{object of the class \command{"dyntermstrc_yields"}.}
  \item{\dots}{further arguments.}
}
