\name{plot.zeroyields}
\Rdversion{1.1}
\alias{plot.zeroyields}

\title{S3 Plot Method
}
\description{S3 plot method for objects of the class
  \command{"zeroyields"}.  The method plots the estimated three-dimensional spot
rate curve. 
}
\usage{
\method{plot}{zeroyields}(x, ...)
}

\arguments{
  \item{x}{object of the class \command{"zeroyields"}
}
  \item{\dots}{further arguments
}
}
