\name{print.dyncouponbonds}
\alias{print.dyncouponbonds}

\title{S3 Print Method
}
\description{The method prints basic information of a dynamic bond data set.
}
\usage{
\method{print}{dyncouponbonds}(x, ...)
}

\arguments{
  \item{x}{object of the class \command{"dyncouponbonds"}
}
  \item{\dots}{further arguments
}
}

