\name{print.dyntermstrc_nss}
\Rdversion{1.1}
\alias{print.dyntermstrc_nss}

\title{S3 Print Method}
\description{S3 print method for objects of the class
  \command{"dyntermstrc_nss"}. The basic parameter and a
  summary of the estimated term structure parameters are printed.}
\usage{
\method{print}{dyntermstrc_nss}(x, ...)
}

\arguments{
  \item{x}{object of the class \command{"dyntermstrc_nss"}}
  \item{\dots}{further arguments}
}
