\name{print.dyntermstrc_yields}
\Rdversion{1.1}
\alias{print.dyntermstrc_yields}

\title{S3 Print Method
}
\description{S3 print method for objects of the class
  \command{"dyntermstrc_yields"}. The method prints information from the
term structure estimation and a summary of the estimated parameters.}
\usage{
\method{print}{dyntermstrc_yields}(x, ...)
}

\arguments{
  \item{x}{object of the class \command{"dyntermstrc_yields"}
}
  \item{\dots}{further arguments}
}
