\name{print.nelson}
\alias{print.nelson}

\title{  S3 Print Method  }

\description{
 S3 print method for an object of the class \command{"nelson"}.
}
\usage{
\method{print}{nelson}(x,...)
}
\arguments{
  \item{x}{object of the class \command{"nelson"}}
  \item{\dots}{other arguments.}
}
\details{
 The print method for an object of the class \command{"nelson"} prints
 important input parameters of the optimisation and the results of the optimisation (i.e. the optimal parameter vector). 
}


%\value{}

%\references{ }
%\author{Robert Ferstl, Josef Hayden}
%\note{ }

\seealso{\command{\link{plot.nelson}}, \command{\link{summary.nelson}} }

%\examples{}

