\name{print.summary.cubicsplines}
\alias{print.summary.cubicsplines}

\title{S3 Print Method }
\description{
  S3 print method for an object of the class \code{summary.cubicsplines}
}
\usage{
\method{print}{summary.cubicsplines}(x, ...)
}

\arguments{
   \item{x}{object of the class \code{summary.nelson}}
   \item{\dots}{other arguments}
}
%\details{}

%\value{}
%\references{}
\author{Robert Ferstl, Josef Hayden }
%\note{}
\seealso{\code{\link{print.summary.nelson}} }
%\examples{}
\keyword{ }

