\name{print.summary.dyntermstrc_nss}
\Rdversion{1.1}
\alias{print.summary.dyntermstrc_nss}

\title{S3 Print Method}
\description{Print method for objects of the class \command{"summary.dyntermstrc_nss"}
}
\usage{
\method{print}{summary.dyntermstrc_nss}(x, ...)
}

\arguments{
  \item{x}{object of the class \command{"summary.dyntermstrc"}}
  \item{\dots}{further arguments
}
}
