\name{print.summary.dyntermstrc_param}
\Rdversion{1.1}
\alias{print.summary.dyntermstrc_param}

\title{S3 Print Method}
\description{
S3 print method for an object of the class \command{"summary.dyntermstrc_param"}
}
\usage{
\method{print}{summary.dyntermstrc_param}(x, ...)
}

\arguments{
  \item{x}{object of the class \command{"summary.dyntermstrc_param"}}
  \item{\dots}{further arguments}
}



