\name{print.summary.dyntermstrc_yields}
\Rdversion{1.1}
\alias{print.summary.dyntermstrc_yields}

\title{S3 Print Method}
\description{ S3 print method for objects of the class \command{"summary.dyntermstrc_yields"}.
}
\usage{
\method{print}{summary.dyntermstrc_yields}(x, ...)
}

\arguments{
  \item{x}{object of the class \command{"summary.dyntermstrc_yields"}
}
  \item{\dots}{further arguments.
}
}
