\name{print.summary.nelson}
\alias{print.summary.nelson}

\title{S3 Print Method }
\description{
  S3 print method for an object of the class \command{"summary.nelson"}.
}
\usage{
\method{print}{summary.nelson}(x, ...)
}

\arguments{
   \item{x}{object of the class \command{"summary.nelson"}.}
   \item{\dots}{other arguments.}
}
%\details{}
%\value{}
%\references{  }
%\author{Robert Ferstl, Josef Hayden}
%\note{}
\seealso{ \command{\link{print.summary.cubicsplines}}}
%\examples{}


