\name{print.summary.termstrc_cs}
\Rdversion{1.1}
\alias{print.summary.termstrc_cs}

\title{S3 Print Method }
\description{
  S3 print method for an object of the class \command{"summary.termstrc_cs"}.
}
\usage{
\method{print}{summary.termstrc_cs}(x, ...)
}

\arguments{
   \item{x}{object of the class \command{"summary.termstrc_cs"}.}
   \item{\dots}{other arguments.}
}


