\name{print.summary.termstrc_nss}
\Rdversion{1.1}
\alias{print.summary.termstrc_nss}


\title{S3 Print Method }
\description{
  S3 print method for an object of the class \command{"summary.termstrc_nss"}.
}
\usage{
\method{print}{summary.termstrc_nss}(x, ...)
}

\arguments{
   \item{x}{object of the class \command{"summary.termstrc_nss"}.}
   \item{\dots}{other arguments.}
}
