\name{print.termstrc_cs}
\Rdversion{1.1}
\alias{print.termstrc_cs}

\title{S3 Print Method for termstrc_cs}

\description{
 S3 print method for an object of the class \command{"termstrc_cs"}.
}
\usage{
\method{print}{termstrc_cs}(x,...)
}
\arguments{
  \item{x}{object of the class \command{"termstrc_cs"}.}
  \item{\dots}{other arguments.}  
}
\details{
 The print method for an object of the class \command{"termstrc_cs"} prints the parameter estimates and the associated (robust) standard errors of the cubic spline functions. 
}

\seealso{\command{\link{plot.termstrc_cs}}, \command{\link{summary.termstrc_cs}} }

