\name{print.termstrc_nss}
\Rdversion{1.1}
\alias{print.termstrc_nss}

\title{ S3 Print Method}
\description{Print method for objects of the class
  \command{"termstrc_nss"}. }
\usage{
\method{print}{termstrc_nss}(x, ...)
}

\arguments{
  \item{x}{objects of the class \command{"termstrc_nss"}
}
  \item{\dots}{further arguments
}
}
\details{
The print method for an object of the class \command{"nelson"} prints
 important input parameters of the optimisation and the results (the optimal parameter vector)
}
