\name{print.termstrc_singlecurve}
\alias{print.termstrc_singlecurve}

\title{  S3 print method  }

\description{
 S3 print method 
}
\usage{
print.termstrc_singlecurve(x,...)
}
\arguments{
  \item{x}{ object of the class \code{termstrc_singelcurve}}
  \item{...}{optional arguments}
}
\details{
 The print method for an object of fhe class \code{termstrc_singlecurve} prints
 the used parameters for the optimisation and the calculated optimal parameter set. 
}


%\value{}

%\references{ ~put references to the literature/web site here ~ }
\author{Robert Ferstl, Josef Hayden}
%\note{ }

\seealso{\code{\link{plot.termstrc_singlecurve}},\code{\link{summary.termstrc_singlecurve}} }

%\examples{}

\keyword{ }
