\name{print.zeroyields}
\Rdversion{1.1}
\alias{print.zeroyields}

\title{S3 Print Method}
\description{S3 print method for objects of the class
  \command{"zeroyields"}. The method prints basic information of a
  zeroyields data set.}
\usage{
\method{print}{zeroyields}(x, ...)
}

\arguments{
  \item{x}{object of the class \command{"zeroyields"}
}
  \item{\dots}{further arguments}
}
