\name{rm_bond}
\alias{rm_bond}
\title{ Bonds removal from a list of bonds }
\description{
 Specified bonds and their associated data are removed from the bond list
}
\usage{
rm_bond(bdata, ISIN, gr)
}

\arguments{
  \item{bdata}{the list with bond data }
  \item{ISIN}{the ISIN numbers of the bonds to remove }
  \item{gr}{the group of bonds where the bonds to be removed belong to}
}
%\details{}
\value{bond list}
%\references{  }
\author{ Robert Ferstl, Josef Hayden }
%\note{ }
%\seealso{}
%\examples{}

\keyword{  }
