\name{spotrates}
\alias{spotrates}

\title{Function for the Calculation of the Spot Rates }
\description{
 The function calculates the spot rates for the chosen approach, a provided maturity vector and
 parameter set
}

\usage{
spotrates(method, beta, m)
}



\arguments{
  \item{method}{\code{"Nelson/Siegel"},\code{"Svensson"}}
  \item{beta}{ parameter set  }
  \item{m}{maturity vector }
}
%\details{}

%\value{}

%\references{ ~put references to the literature/web site here ~ }
\author{Robert Ferstl, Josef Hayden }

%\note{ ~~further notes~~ }
%\seealso{ ~~objects to See Also as \code{\link{help}}, ~~~ }

%\examples{}

\keyword{ }