\name{srates}
\alias{srates}

\title{Function for the calculation of the spot rates }
\description{
 The function calculates the spot rates for the chosen approach, a provided maturity vector and
 parameter set
}

\usage{
srates(method, beta, m)
}



\arguments{
  \item{method}{\code{"Nelson/Siegel"},\code{"Svensson"}}
  \item{beta}{ paramater set  }
  \item{m}{maturity vector }
}
%\details{}

%\value{}

%\references{ ~put references to the literature/web site here ~ }
\author{Robert Ferstl, Josef Hayden }

%\note{ ~~further notes~~ }
%\seealso{ ~~objects to See Also as \code{\link{help}}, ~~~ }

%\examples{}

\keyword{ }

