\name{summary.cubicsplines}
\alias{summary.cubicsplines}

\title{ S3 summary method }
\description{
  S3 summary method for objects of the class \code{cubicsplines}
}
\usage{
summary.cubicsplines(x,...)
}

\arguments{
   \item{x}{object of the class \code{cubicsplines} }
   \item{...}{optional arguments}
}
\details{
  The summary method for an object of the class \code{nelson} prints the solution of
 the goodness of fit tests (RMSE, AABSE) of the optimisation. 
}
%\value{}
%\references{ }
\author{Robert Ferstl, Josef Hayden }
%\note{}
\seealso{\code{\link{plot.cubicsplines}},\code{\link{print.cubicsplines}}, \code{\link{rmse}},\code{\link{aabse}}}
\examples{}

\keyword{ }

