\name{summary.dyntermstrc_yields}
\Rdversion{1.1}
\alias{summary.dyntermstrc_yields}

\title{S3 Summary Method
}
\description{Summary method for objects of the class
  \command{"dyntermstrc_yields"}. The mean RMSE and AABSE of the yields
  is calculated
}
\usage{
\method{summary}{dyntermstrc_yields}(object, ...)
}

\arguments{
  \item{object}{object of the class \command{"dyntermstrc_yields"}
}
  \item{\dots}{further arguments
}
}
