\name{summary.termstrc_singlecurve}
\alias{summary.termstrc_singlecurve}

\title{ S3 summary method  }
\description{S3 summary method}
\usage{
summary.termstrc_singlecurve(x)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{object of the class \code{termstrc_singlecurve} }
}
\details{
 The summary method for an object of the class \code{termstrc_singlecurve} prints the solution of
 the goodness of fit tests for the yields and prices. Moreover a convergence information of the used
 optimiser (nlminb) is printed. 
}


%\value{}


%\references{ ~put references to the literature/web site here ~ }


\author{Robert Ferstl, Josef Hayden}

%\note{ }

\seealso{ \code{\link{nlminb}},\code{\link{plot.termstrc_singlecurve}},\code{\link{print.termstrc_singlecurve}}}


%\examples{}

\keyword{}