\name{summary.zeroyields}
\Rdversion{1.1}
\alias{summary.zeroyields}

\title{S3 Summary Method}
\description{S3 summary method for objects of the class
  \command{"zeroyields"}. The method calculates basic summary statistics
of the data.}
\usage{
\method{summary}{zeroyields}(object, ...)
}

\arguments{
  \item{object}{object of the class \command{"zeroyields"}
}
  \item{\dots}{further arguments
}
}

\value{returns an object of the class \command{"summary.zeroyields"}
}
