\name{termstrc-package}
\alias{termstrc-package}
\alias{termstrc}
\docType{package}


\title{
Term Structure and Credit Spread Estimation
}


\description{
The package includes two widely-used term structure estimation procedures, 
the parametric Nelson and Siegel approach and the Svensson approach.
Moreover the traditional way of credit spread calculation is implemented. 
}


\details{
\tabular{ll}{
Package: \tab termstrc\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2006-11-27\cr
License: \tab GPL \cr
}
}


\author{Robert Ferstl, Josef Hayden

Maintainer: Robert Ferstl <robert.ferstl@wu-wien.ac.at>

Special thanks to the Finance Management Science Lab 2005 

}


\references{
 Charles R. Nelson and Andrew F. Siegel. Parsimonious modeling of yield curves. The Journal 
of Business, 60(4):473Ð489, Oct 1987. 

Lars E.O. Svensson. Estimating and interpreting forward interest rates: Sweden 1992 -1994. 
Technical Reports 4871, National Bureau of Economic Research, Inc, Sep 1994.

Yield Curve Modelling at the Bank of Canda, David Bolder and David Streliski, Technical Report No 84 Bank of Canada 1999.

}

\keyword{ package }
