\name{zyields}
\alias{zyields}
\docType{data}
\title{Zero Coupon Yield Data Set}
\description{Zero Coupon Yield Data Set}

\usage{data(zyields)}

\note{
If you use your own data set, make sure that the structure is identical to the provided data set.
Use the function \command{str()} to explore the data set.}


\keyword{datasets}
