\name{eurobonds}
\alias{eurobonds}
\docType{data}
\title{European Government Bonds}
\description{European government bonds}

\usage{data(eurobonds)}

\details{
The data set eurobonds consists of German, Austrian, Italian and Hungarian government bonds.
}

\note{
If you use your own dataset, make sure that the structure is identical to the provided data sets.
Use the function \code{str()} to explore the data set.}


\seealso{\code{\link{corpbonds}}}

\examples{
data(corpbonds)
str(corpbonds) 
}

\keyword{datasets}
