\name{fwr_sv}
\alias{fwr_sv}

\title{Forward rate calculation according to Svensson}
\description{
Calculate forward rates according to the Svensson approach. 
}
\usage{
fwr_sv(beta, m)
}

\arguments{
  \item{beta}{Parameter vector }
  \item{m}{Maturity vector }
}
%\details{}
%\value{}


%\references{  }
\author{ Robert Ferstl, Josef Hayden }
%\note{ }
\seealso{\code{\link{fwr_ns}}, \code{\link{forwardrates}}}
%\examples{}

\keyword{  }
