\name{gi}
\alias{gi}

\title{ Cubic Functions }
\description{
  Calculation of the cubic functions according to an approach of
  McCulloch (1975)
}
\usage{gi(t, T, i, s)}

\arguments{
  \item{t}{maturity }
  \item{T}{knot points}
  \item{i}{index}
  \item{s}{number of basis functions}}
}
%\details{...}
%\value{
% \item{g}{..g...}
% }
\references{ 

J. Huston McCulloch (1971):
Measuring the Term Structure of Interest Rates.
\emph{The Journal of Business, \bold{44} 19--31.} 


J. Huston McCulloch (1975):
The Tax-Adjusted Yield Curve.
\emph{The Journal of Finance, \bold{30} 811--830.} 


Sanjay K. Nawalkha and Gloria M. Soto and Natalia K. Beliaeva (2005):
Interest Rate Risk Modeling : The Fixed Income Valuation Course
\emph{Wiley Finance,60--67}

 }
\author{Robert Ferstl, Josef Hayden }
%\note{ }
%\seealso{  }
%\examples{e.g. example from the textbook }

\keyword{ }

