\name{gi}
\Rdversion{1.1}
\alias{gi}

\title{ Cubic Functions }
\description{
  Calculation of the cubic functions according to the approach of
  McCulloch (1975).
}
\usage{gi(t, T, i, s)}

\arguments{
  \item{t}{maturity. }
  \item{T}{knot points.}
  \item{i}{index.}
  \item{s}{number of basis functions.}}

\references{ 
J. Huston McCulloch (1975):
The Tax-Adjusted Yield Curve.
\emph{The Journal of Finance, \bold{30} 811--830.} 

 }
