\name{govbonds}
\alias{govbonds}
\docType{data}
\title{European Government Bonds}
\description{European government bonds.}

\usage{data(eurobonds)}

\details{
The data set \command{bonds} consists of German, Austrian, French, Belgian, Finnish and Spanish,  government bonds obtained from Thomson Datastream.
}

\note{
If you use your own data set, make sure that the structure is identical to the provided data sets.
Use the function \command{str()} to explore the data set.}


\seealso{\command{\link{eurobonds}}, \command{\link{corpbonds}}}

\examples{
data(govbonds)
str(govbonds) 
}

\keyword{datasets}
