\name{govbonds}
\alias{govbonds}
\docType{data}
\title{European Government Bonds}
\description{European government bonds.}

\usage{data(eurobonds)}

\details{
The data set \code{bonds} consists of German, Austrian, French, Belgian, Finnish and Spanish,  government bonds obtained from Datastream.
}

\note{
If you use your own dataset, make sure that the structure is identical to the provided data sets.
Use the function \code{str()} to explore the data set.}


\seealso{\code{\link{eurobonds}}}

\examples{
data(govbonds)
str(govbonds) 
}

\keyword{datasets}
