\name{print.cubicsplines}
\alias{print.cubicsplines}

\title{S3 Print Method for Cubicsplines}

\description{
 S3 print method for an object of the class \code{"cubicsplines"}.
}
\usage{
\method{print}{cubicsplines}(x,...)
}
\arguments{
  \item{x}{object of the class \code{"cubicsplines"}.}
  \item{\dots}{other arguments.}  
}
\details{
 The print method for an object of the class \code{"cubicsplines"} prints the results of the regression analysis. (i.e. the parameter of the cubic spline functions). 
}


%\value{}

%\references{ }
%\author{Robert Ferstl, Josef Hayden}
%\note{ }

\seealso{\code{\link{plot.cubicsplines}}, \code{\link{summary.cubicsplines}} }

%\examples{}

\keyword{ }
