\name{rm_bond}
\alias{rm_bond}
\title{Remove Bonds from a Dataset }
\description{
 Specified bonds and their associated data are removed from the bond data set.
}
\usage{
rm_bond(bdata, ISIN, gr)
}

\arguments{
  \item{bdata}{bond data set. }
  \item{ISIN}{the ISIN numbers of the bonds to remove. }
  \item{gr}{the group  where the bonds to be removed belong to.}
}
%\details{}
\value{The function returns the new bond data set.}
%\references{  }
%\author{ Robert Ferstl, Josef Hayden }
%\note{ }
%\seealso{}
%\examples{}


