\name{termstrc-package}
\alias{termstrc-package}
\alias{termstrc}
\docType{package}


\title{
Term Structure and Credit Spread Estimation
}


\description{
The package offers several widely-used term structure estimation procedures,
 i.e. the parametric Nelson and Siegel approach, Svensson approach and cubic splines.
}

\references{

Bank for International Settlements (2005).
Zero-coupon yield curves: technical documentation.
\emph{BIS Papers, \bold{No. 25}}


Robert R. Bliss (2007):
Testing term structure estimation methods.
\emph{Advances in Futures and Options Research, \bold{9} 197--232.} 


David Bolder and David Streliski (1999):
Yield Curve Modelling at the Bank of Canada.
\emph{Bank of Canada Technical Report, \bold{No. 84} }


Alois Geyer and Richard Mader (1999):
Estimation of the Term Structure of Interest Rates - A Parametric Approach.
\emph{OeNB Working Paper, \bold{No. 37}}


Michalis Ioannides (2003):
A comparison of yield curve estimation techniques using UK data.
\emph{Journal of Banking \& Finance, \bold{27} 1--26.} 


J. Huston McCulloch (1971):
Measuring the Term Structure of Interest Rates.
\emph{The Journal of Business, \bold{44} 19--31.} 

J. Huston McCulloch (1975):
The Tax-Adjusted Yield Curve.
\emph{The Journal of Finance, \bold{30} 811--830.} 


Sanjay K. Nawalkha and Gloria M. Soto and Natalia K. Beliaeva (2005):
Interest Rate Risk Modeling : The Fixed Income Valuation Course
\emph{Wiley Finance}

Charles R. Nelson and Andrew F. Siegel (1987): 
Parsimonious modeling of yield curves.
\emph{The Journal of Business}, \bold{60(4)}:473--489.


Lars E.O. Svensson (1994): 
Estimating and interpreting forward interest rates: Sweden 1992 -1994. 
\emph{Technical Reports 4871, National Bureau of Economic Research}. 
 
}

\keyword{package}
