\name{termstrc-package}
\alias{termstrc-package}
\alias{termstrc}
\docType{package}


\title{
Term Structure and Credit Spread Estimation
}


\description{
The package offers functions for the estimation of zero-coupon yield (spread) curves 
based on three widely-used term structure estimation procedeures, i.e. the
parametrice Nelson and Siegel , the Svensson (1994) approach and the cubic splines approach of
J.Huston and McCulloch (1971).  
}

\references{

David Bolder and David Streliski (1999):
Yield Curve Modelling at the Bank of Canada.
\emph{Technical Report No 84 Bank of Canada} 

J.Huston and McCulloch (1971):
Measuring the Term Structure of Interest Rates.
\emph{The Journal of Business, \bold{44} 19--31.} 

Charles R. Nelson and Andrew F. Siegel (1987): 
Parsimonious modeling of yield curves.
\emph{The Journal of Business}, \bold{60(4)}:473--489.

Lars E.O. Svensson (1994): 
Estimating and interpreting forward interest rates: Sweden 1992 -1994. 
\emph{Technical Reports 4871, National Bureau of Economic Research}. 
 
}

\keyword{package}
